Adding Switch & Track
Switch and Track both have good-enough comments in their respective files to figure out what exactly they do. They are NOT yet added to the main game - there is a change that creates tracks out of them, but for now it is commented out. 

Switch runs on a loop, and can handle more than 1 branching off. I'll add a function to do a actually activate the switch from the track later on. If anyone wants to do it now because they hate themselves, define a function in Track that just calls self.switch.activate().