Levi's parts are done
Images instead of circles are now blitted to the screen.
Next object coming up has been implemented.
The score and directions are in better locations
Had to fix end point detection. There are only 4 end points at the ends
of tracks 4, 6, 9, and 10 (bottle, glass bottle, paper, and trash
respectively)
Added a helpful little map showing which tracks map to which ids.
Score increases/decreases and is displayed correctly.
The DIN Condensed font I found is close, but not quite the same as the
background. Hardly noticeable, honestly.